requirements.txt
asgiref==3.2.10
attrs==20.2.0
bcrypt==3.2.0
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coverage==5.3
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.2
django-bootstrap-modal-forms==2.0.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-sortable-listview==0.43
django-widget-tweaks==1.4.8
gunicorn==20.0.4
html5lib==1.1
idna==2.10
iniconfig==1.1.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.10.0
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.1
urllib3==1.25.11
WeasyPrint==52
webencodings==0.5.1
